+++ spec/lib/http_stub/rake/server_daemon_tasks_smoke_spec.rb
@@ -16,7 +16,7 @@ describe HttpStub::Rake::ServerDaemonTasks do
       end
 
       it "starts a stub server that responds to stub requests" do
-        request = Net::HTTP::Post.new("/stubs")
+        request = Net::HTTP::Post.new("/http_stub/stubs")
         request.body = { response: { status: 302, body: "Some Body" } }.to_json
 
         response = Net::HTTP.new("localhost", 8002).start { |http| http.request(request) }